pyproject.toml
[project]
name = "carnet-tbi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.97.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pdfplumber>=0.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
